+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.2] - 2021-11-08
+
+**Added:**
+- Changelog version extract command
+- Cert reading as C-string command
+- Batch suffix command
+
 ## [0.0.1] - 2021-11-06
 
 **Added:**